Move component folder into _ui folder
