+++ src/app/_ui/components/FounderVision/FounderVision.module.css
