+++ src/app/_ui/components/FounderVision/FounderVision.tsx
