+++ src/app/_ui/components/TextSlider/TextSlider.module.css
