+++ src/app/_ui/components/TextSlider/TextSlider.tsx
