+++ src/app/_ui/components/TextSlider/TextSliderWrapper.tsx
