+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import { outfit, raleway } from './_ui/fonts';
 import './_ui/token.css';
 import './_ui/globals.css';
-import { HeaderWrapper } from './_components/Header/HeaderWrapper';
-import { Footer } from './_components/Footer/Footer';
+import { HeaderWrapper } from './_ui/components/Header/HeaderWrapper';
+import { Footer } from './_ui/components/Footer/Footer';
 
 
 export const metadata: Metadata = {
