+++ src/app/page.tsx
@@ -1,8 +1,8 @@
-import { BenefitList } from './_components/BenefitList/BenefitList';
-import { FounderVision } from './_components/FounderVision/FounderVision';
-import { TextSliderWrapper } from './_components/TextSlider/TextSliderWrapper';
-import { Tour } from './_components/Tour/Tour';
-import styles from './page.module.css'
+import { BenefitList } from './_ui/components/BenefitList/BenefitList';
+import { FounderVision } from './_ui/components/FounderVision/FounderVision';
+import { TextSliderWrapper } from './_ui/components/TextSlider/TextSliderWrapper';
+import { Tour } from './_ui/components/Tour/Tour';
+import styles from './page.module.css';
 
 export default function Home() {
   return (
